rename datasets sampling shared logic file to utils.py
